Move boto-needing AWS imports to a new file
