+++ src/toil/batchSystems/awsBatch.py
@@ -48,7 +48,8 @@ from toil.batchSystems.abstractBatchSystem import (EXIT_STATUS_UNAVAILABLE_VALUE
 from toil.batchSystems.cleanup_support import BatchSystemCleanupSupport
 from toil.common import Config, Toil
 from toil.job import JobDescription
-from toil.lib.aws import establish_boto3_session, get_current_aws_zone, zone_to_region
+from toil.lib.aws import get_current_aws_zone, zone_to_region
+from toil.lib.aws.session import establish_boto3_session
 from toil.lib.conversions import b_to_mib, mib_to_b
 from toil.lib.misc import slow_down, utc_now, unix_now_ms
 from toil.lib.retry import retry
