+++ src/toil/test/lib/test_ec2.py
@@ -19,7 +19,7 @@ from toil.lib.aws.ami import (
     get_flatcar_ami,
     official_flatcar_ami_release,
 )
-from toil.lib.aws import establish_boto3_session
+from toil.lib.aws.session import establish_boto3_session
 from toil.test import ToilTest, needs_aws_ec2
 
 logger = logging.getLogger(__name__)
